refactor int->uint8_t/int8_t

+++ dht11sensor.cpp
@@ -7,7 +7,7 @@
 
 
 
-DHT11Sensor::DHT11Sensor(String _name, int _pin)
+DHT11Sensor::DHT11Sensor(String _name, uint8_t _pin)
   : Sensor::Sensor()
   , DHT::DHT(_pin, DHTTYPE)
 {
@@ -41,6 +41,6 @@ String* DHT11Sensor::getDescription(){
   return this->description;
 }
 
-int DHT11Sensor::getNrOfValues(){
+uint8_t DHT11Sensor::getNrOfValues(){
   return this->nrOfValues;
 }
